Added Vertical/Horizontal Output Mixer for Gyro
since our general Gyro Sens is by Natural Sensitivity Scale values (except for Crosshair portion), we're taking a different approach by adding V/H (in this case: X/Y) Output Mixer setup. This affects both Gyro Camera and Crosshair.

Moving the slider to the left reduces vertical sensitiviity (which is X), Moving the slider to the right horizontal sensitiviity (which is Y). theoretically: you can set the slider to -43% and have the Gyro Ratio to be 16:9 Gyro ratio. By default: 0% is 1:1 Ratio.