initial support for Gyro Aiming on Splitscreen Multiplayer
Revised overall gyro code to be based on a per-controller basis, allowing you to have a local gyro multiplayer session.  However: the Crosshair Aiming portion for Player 2/3/4 does not work at the moment.

this also improves the crosshair edge turning for gyro, making it a bit consistent with mouse input's