Gyro Calibration UI added to Gyro Menu
Added the ability to manually calibrate the controller's Motion Sensors within the Menu UI, no need to press the Calibration button! Just place the controller on a surface, and wait for 5 seconds, giving you ample time to put the controller on the floor, as the calibration is instant! 

If you accidentally move the controller during the gyro calibration process: press the ACCEPT key two times inorder to restart the Gyro Calibration process!