Vibration Filter support
added `RumbleFilter` function to reduce the impact of lower-frequency rumbles (such as DualShock 4 controllers) by roughly 70% percent, best use for Gyro Aiming or Mixed Input.

This option is only accessible within `pd.ini`. By default: `-1` will be enabled for when Gyro Aim support gets added, in the meantime: it'll be treated as `0`. However: it can still be forcefully enabled by setting it to `1`.