Joystick Crosshair now uses Crosshair Edge Boundary slider
This wasn't explained properly, but that slider was meant for Mouse Input, but now Joystick obtain the same benefit. But to ensure the migration remains smooth: the joystick boundary slider is slightly higher than Mouse inorder to maintain the same crosshair boundary as before (if set to 60%).

In case you didn't noticed: `crosshairedgeboundary` default has been changged to `6` to match the same setup as the joystick's, while changing the slider to be a percentage instead of a simple "0 to 10"